+++ R/mathml.R
@@ -53,6 +53,9 @@ mathml <- function(term=quote((a + b)^2L == a^2L + 2L*a*b + b^2L), flags=NULL,
   t <- rolog::once(call("r2mathml", term, expression(X), flags),
     options=list(preproc=list(rolog::preproc, mathml_preproc)),
     env=env)
+  if(!is.list(t))
+    stop("r2mathml failed: ", as.character(term), " return value: ", t)
+  
   r <- paste(t$X, collapse="")
   if(flags$cat)
     return(cat(r))
@@ -360,16 +363,19 @@ canonical <- function(term=quote(`%in%`(table=Table, x=X)), drop=TRUE)
 
 #' Hook for custom symbols
 #' 
+#' hook(function)
 #' hook(term, display)
 #' hook_fn(fn)
 #' unhook(term)
 #' hooked(term)
 #'
 #' @param term
-#' an R call or symbol/number. This is the expression to replace.
+#' an R call or symbol/number or a function. This is the expression 
+#' to replace.
 #'
-#' @param display
-#' an R call or symbol/number. This is shown instead of _term_.
+#' @param display (default is NULL)
+#' an R call or symbol/number. This is shown instead of _term_. If _term_ is
+#' a function, display is NULL (the function body is rendered instead).
 #'
 #' @param quote (default is TRUE)
 #' indicates that _term_ and _display_ should be quoted.
@@ -387,6 +393,7 @@ canonical <- function(term=quote(`%in%`(table=Table, x=X)), drop=TRUE)
 #' @md
 #'
 #' @examples
+#' hook(dot)
 #' hook(t0, subscript(t, 0))
 #' hooked(quote(t0))
 #' mathml(quote(t0))
@@ -400,17 +407,35 @@ hook <- function(term, display=NULL, quote=TRUE, as.rolog=TRUE)
 {
   if(quote)
   {
-    term <- substitute(term)
-    display <- substitute(display)
+    term1 <- substitute(term)
+    display1 <- substitute(display)
+  }
+  else
+  {
+    term1 <- term
+    display1 <- display
+  }
+  
+  if(is.null(display1) && is.function(term))
+  {
+    name <- as.character(term1)
+    fm <- formals(term)
+    args <- names(fm)
+    dotargs <- sprintf(".%s", args)
+    subst <- lapply(dotargs, as.name)
+    arg1 <- as.call(c(as.name(name), subst))
+    names(subst) <- args
+    arg2 <- do.call(substitute, list(body(term), subst))
+    return(hook(arg1, arg2, quote=FALSE))
   }
   
   if(as.rolog)
   {
-    term <- rolog::as.rolog(term)
-    display <- rolog::as.rolog(display)
+    term1 <- rolog::as.rolog(term1)
+    display1 <- rolog::as.rolog(display1)
   }
 
-  r <- rolog::once(call("asserta", call("math_hook", term, display)))
+  r <- rolog::once(call("asserta", call("math_hook", term1, display1)))
   if(isFALSE(r))
     return(FALSE)